Reorganization and bugfix
